backend/requirements.txt
# Backend Dependencies for Smart Farming Platform
# Updated for Python 3.13 compatibility

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.0

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25

# Authentication
PyJWT==2.8.0

# Machine Learning (Python 3.13 compatible)
scikit-learn==1.5.2
tensorflow>=2.17.0

# Data Processing (Python 3.13 compatible)
numpy>=1.26.0,<2.2.0
pandas>=2.2.0
scipy>=1.13.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.9.0

# Model Persistence
joblib>=1.3.0

# Visualization (for training/evaluation)
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Data Augmentation
imgaug>=0.4.0
albumentations>=1.3.0
